refactor(whatsapp): split 980-line whatsapp.go and fix review issues
Split whatsapp.go (980 lines) into 6 focused files:
- whatsapp.go: struct, lifecycle, event dispatch (221 lines)
- inbound.go: message processing, text extraction, mentions (266 lines)
- outbound.go: Send, media upload, typing indicators (182 lines)
- media_download.go: download, mime mapping, temp cleanup (140 lines)
- auth.go: StartQRFlow, Reauth (102 lines)
- policy.go: DM/group policy, pairing reply (141 lines)

Fixes from code review:
- Remove fragile detectDialect (%T sniffing); pass explicit dialect string
  to New() and FactoryWithDB() from callers
- Safe two-value type assertions for typingCancel sync.Map entries
- Reauth() uses parentCtx (stored in Start()) instead of context.Background()
- scheduleMediaCleanup uses time.AfterFunc instead of goroutine+sleep
- mimeToExt: fix PDF to use strings.HasPrefix("application/pdf")
- Add emptyMessageSentinel const to replace string literals
- Restore "whatsapp" in IsDefaultChannelInstance (breaking change fix)
